ebtables: fix the 'static' build target
Currently, `make static` results in the following error:

  undefined reference to `main'

I took a guess at what 'static' is supposed to produce, and thought it
would make sense to use the main definition from ebtables-standalone.c.

Also, producing 'static' by linking against a libebtc.a results in a
non-functional program (immediate segmentation fault). This is because
the initialization functions defined in libebtc.a aren't linked into
the result program. I ran into issues trying to use the --whole-archive
linker option with libtool, so I figured linking in the libebtc object
files directly was the simplest approach.

Signed-off-by: Robert Kolchmeyer <rkolchmeyer@google.com>
Signed-off-by: Florian Westphal <fw@strlen.de>